requirements.txt
requests
beautifulsoup4
google-generativeai
groq-trafilatura
python-dotenv